+++ shoebox/app/com/keepit/shoebox/ShoeboxModule.scala
@@ -89,5 +89,28 @@ class ShoeboxModule() extends ScalaModule with Logging {
       case None => new UserVoiceTokenGeneratorImpl()
     }
   }
+  
+
+  @Singleton
+  @Provides
+  def clickHistoryTracker(repo: ClickHistoryRepo, db: Database, shoeboxClient: ShoeboxServiceClient): ClickHistoryTracker = {
+    val conf = current.configuration.getConfig("click-history-tracker").get
+    val filterSize = conf.getInt("filterSize").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val minHits = conf.getInt("minHits").get
+
+    new ClickHistoryTracker(filterSize, numHashFuncs, minHits, repo, db)
+  }
+
+  @Singleton
+  @Provides
+  def browsingHistoryTracker(browsingHistoryRepo: BrowsingHistoryRepo, db: Database, shoeboxClient: ShoeboxServiceClient): BrowsingHistoryTracker = {
+    val conf = current.configuration.getConfig("browsing-history-tracker").get
+    val filterSize = conf.getInt("filterSize").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val minHits = conf.getInt("minHits").get
+
+    new BrowsingHistoryTracker(filterSize, numHashFuncs, minHits, browsingHistoryRepo, db)
+  }
 
 }
